+++ cmd/neofs-node/config/validate_test.go
@@ -191,7 +191,7 @@ storage:
     small_object_size: 100 kb
 
     blobstor:
-    - perm: 0644
+      perm: 0644
       depth: 5
 
     gc:
@@ -213,7 +213,7 @@ storage:
         max_batch_delay: 10ms
 
       blobstor:
-      - type: fstree
+        type: fstree
         path: tmp/0/blob
 `,
 			wantErr: false,