+++ pkg/local_object_storage/blobstor/info.go
@@ -5,13 +5,8 @@ func (b *BlobStor) DumpInfo() Info {
 	b.modeMtx.RLock()
 	defer b.modeMtx.RUnlock()
 
-	sub := make([]SubStorageInfo, len(b.storage))
-	for i := range b.storage {
-		sub[i].Path = b.storage[i].Storage.Path()
-		sub[i].Type = b.storage[i].Storage.Type()
-	}
-
 	return Info{
-		SubStorages: sub,
+		Path: b.storage.Storage.Path(),
+		Type: b.storage.Storage.Type(),
 	}
 }